+++ ntex/src/web/ws.rs
@@ -209,7 +209,7 @@ where
                     description: None,
                 }))))
             }
-            DispatchItem::Stop(Reason::KeepAliveTimeout) => Err(WsError::KeepAlive),
+            DispatchItem::Stop(Reason::KeepAlive) => Err(WsError::KeepAlive),
             DispatchItem::Stop(Reason::ReadTimeout) => Err(WsError::ReadTimeout),
             DispatchItem::Stop(Reason::WriteTimeout) => Err(WsError::WriteTimeout),
             DispatchItem::Stop(Reason::Decoder(e)) => {